#89 ad spaces/panel initialization fixes
git-svn-id: file:///Users/michaelkoepke/svn/semiologic/themes/semiologic/trunk@655 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ inc/upgrade/options.php
@@ -87,7 +87,9 @@ function sem_clean_up_options()
 	
 	# drop obsolete tables
 	$wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}WP_HASHCASH;");
-	$wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}sem_ad_spaces;");
+	
+	if ( !in_array('sem-ad-space/sem-ad-space.php', $active_plugins) )
+		$wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}sem_ad_spaces;");
 } # clean_up_options()
 
 add_action('shutdown', 'sem_clean_up_options');
